Reducers and LocationService refactor
