+++ src/js/actions/types.js
@@ -3,5 +3,5 @@ export const HIDE_SPINNER = 'HIDE_SPINNER';
 
 export const FETCH_REASONCODES = 'FETCH_REASONCODES';
 export const FETCH_USERS = 'FETCH_USERS';
-export const FETCH_CURRENT_LOCATION = 'FETCH_CURRENT_LOCATION';
+export const FETCH_SESSION_INFO = 'FETCH_SESSION_INFO';
 export const CHANGE_CURRENT_LOCATION = 'CHANGE_CURRENT_LOCATION';
