+++ src/js/reducers/index.jsx
@@ -3,14 +3,14 @@ import { localeReducer } from 'react-localize-redux';
 import spinnerReducer from './spinnerReducer';
 import usersReducer from './usersReducer';
 import reasonCodesReducer from './reasonCodesReducer';
-import locationsReducer from './locationsReducer';
+import sessionReducer from './sessionReducer';
 
 const rootReducer = combineReducers({
   locale: localeReducer,
   spinner: spinnerReducer,
   reasonCodes: reasonCodesReducer,
   users: usersReducer,
-  location: locationsReducer,
+  session: sessionReducer,
 });
 
 export default rootReducer;
